fix(gemm): label DTYPE as bf16 to match bfloat16 factories
example_gemm.py's matmul / matmul_nt factories default to T.bfloat16
and bench_one() creates torch.bfloat16 inputs, but the module-level
DTYPE constant still said "fp16". That label is propagated into the
benchmark result dicts ({"dtype": DTYPE}) and into compare_bench.py
keys, so any baseline comparison would silently overlay bf16
measurements onto fp16 baselines under the same key, producing
misleading deltas.

Switch DTYPE to "bf16" to match the actual numeric type.

+++ gemm/example_gemm.py
@@ -103,7 +103,7 @@ def run_regression_perf():
 # ---------------------------------------------------------------------------
 
 OP_NAME = "gemm"
-DTYPE = "fp16"
+DTYPE = "bf16"
 
 _SHAPES = [
     # (M, N, K, block_M, block_N, block_K, num_stages, num_threads)
